Re-think sequence definitions
Make a sequence driven by Operations, with a set of initial operations
and further operations that can be defined as 'Next' actions upon
completion of one operation. This allows for cycles of steps and
sequences that never end.
Add a fluent API for sequence creation rather than having to define the
structure directly.
Improve ease of InterpolateSolid effect creation.